T0424 FEAT: Add field to the tag to help identify them
We are adding some fields to give more information to the tag. The main 3 are the author, the department who use the tag and a description. We add those to all tag and show them in the list view.
To add the department we need to add a new dependence the hr base module.

We add a new value to the smart part of the tag. This new value give the possibility to give a end date to a tag. We had a cron who  archive tag who are older than their valid date.